Implement basic URL checking.
Implement basic sanity checks on the passed in URI - checking if
it conforms to the protocols we can accept. Currently, we use the
assert library to do this. Future improvements would have a more
user-friendly handling of this checking, as a core dump is one way
to exit a program, but it's hardly the best. This may also cause issues
once we start getting a GUI involved in the program.