Tokenise uri into host, protocol, and path.
Added strtok_r() based tokeniser (that actually works) to tokenise the
uri string into the bits we want.

This required updating the makefile standard used to gnu18 not c18, and
I updated the dependencies and added CFLAGS while I was in there.

I also refactored the browser.start() function to now include the SSL
and verbose variables.