Parsing headers.
Headers are parsed and placed into an array of httpHeader structs; each
header consists of the header itself and the value. In the response,
these are seperated by ':', so we have split on this character. The
headers are normalised to lowercase, and the leading character is
removed (this will always be a ' ' character).

Finally, we print the rest of the response - the print statement has
changed from printing the whole response to just printing the HTML code.